Make helper for can_partially_complete
