+++ src/menu/menu_functions.rs
@@ -308,6 +308,39 @@ pub fn replace_in_buffer(value: Option<Suggestion>, editor: &mut Editor) {
     }
 }
 
+/// Helper for `Menu::can_partially_complete`
+pub fn can_partially_complete(values: &[Suggestion], editor: &mut Editor) -> bool {
+    if let (Some(Suggestion { value, span, .. }), Some(index)) = find_common_string(values) {
+        let index = index.min(value.len());
+        let matching = &value[0..index];
+
+        // make sure that the partial completion does not overwrite user entered input
+        let extends_input = matching.starts_with(&editor.get_buffer()[span.start..span.end]);
+
+        if !matching.is_empty() && extends_input {
+            let mut line_buffer = editor.line_buffer().clone();
+            line_buffer.replace_range(span.start..span.end, matching);
+
+            let offset = if matching.len() < (span.end - span.start) {
+                line_buffer
+                    .insertion_point()
+                    .saturating_sub((span.end - span.start) - matching.len())
+            } else {
+                line_buffer.insertion_point() + matching.len() - (span.end - span.start)
+            };
+
+            line_buffer.set_insertion_point(offset);
+            editor.set_line_buffer(line_buffer, UndoBehavior::CreateUndoPoint);
+
+            true
+        } else {
+            false
+        }
+    } else {
+        false
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
